move geo module as website.tools submodule
